09/06/2012 - 1
-Added an introduction to the test specification.
-Fixed some errors in test names and files required.
-Made it look a bit more presentable so that it can just be copied
straight into the final document when it's done.
All my tests are done, I just need to re-test the black box test about a
game ending with a winner and also carry out the tournament test (which
I can't until the winner thing works). Tests that have no result or
status etc are black box tests that are to be carried out by Raj as
these are the acceptance criteria tests he designed. He will also need
to add/update the tables with the tests that he has carried out.